// frontend/src/app/core/model/model.d.ts
export interface ITransaction {

    /** Transaction id */
    id?: (number|null);

    /** Transaction isDeleted */
    isDeleted?: (boolean|null);

    /** Transaction created */
    created: string;

    /** Transaction accountIdFrom */
    accountIdFrom: number;

    /** Transaction accountIdTo */
    accountIdTo: number;

    /** Transaction amountFrom */
    amountFrom: number;

    /** Transaction amountTo */
    amountTo: number;

    /** Transaction categoryId */
    categoryId: number;

    /** Transaction subCategoryId */
    subCategoryId?: (number|null);

    /** Transaction familyMemberId */
    familyMemberId?: (number|null);

    /** Transaction comment */
    comment?: (string|null);

    /** Transaction offBudget */
    offBudget?: (boolean|null);
}

/** Represents a Transaction. */
export class Transaction implements ITransaction {

    /**
     * Constructs a new Transaction.
     * @param [properties] Properties to set
     */
    constructor(properties?: ITransaction);

    /** Transaction id. */
    public id: number;

    /** Transaction isDeleted. */
    public isDeleted: boolean;

    /** Transaction created. */
    public created: string;

    /** Transaction accountIdFrom. */
    public accountIdFrom: number;

    /** Transaction accountIdTo. */
    public accountIdTo: number;

    /** Transaction amountFrom. */
    public amountFrom: number;

    /** Transaction amountTo. */
    public amountTo: number;

    /** Transaction categoryId. */
    public categoryId: number;

    /** Transaction subCategoryId. */
    public subCategoryId: number;

    /** Transaction familyMemberId. */
    public familyMemberId: number;

    /** Transaction comment. */
    public comment: string;

    /** Transaction offBudget. */
    public offBudget: boolean;

    /**
     * Creates a new Transaction instance using the specified properties.
     * @param [properties] Properties to set
     * @returns Transaction instance
     */
    public static create(properties?: ITransaction): Transaction;

    /**
     * Encodes the specified Transaction message. Does not implicitly {@link Transaction.verify|verify} messages.
     * @param message Transaction message or plain object to encode
     * @param [writer] Writer to encode to
     * @returns Writer
     */
    public static encode(message: ITransaction, writer?: $protobuf.Writer): $protobuf.Writer;

    /**
     * Decodes a Transaction message from the specified reader or buffer.
     * @param reader Reader or buffer to decode from
     * @param [length] Message length if known beforehand
     * @returns Transaction
     * @throws {Error} If the payload is not a reader or valid buffer
     * @throws {$protobuf.util.ProtocolError} If required fields are missing
     */
    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): Transaction;

    /**
     * Creates a Transaction message from a plain object. Also converts values to their respective internal types.
     * @param object Plain object
     * @returns Transaction
     */
    public static fromObject(object: { [k: string]: any }): Transaction;

    /**
     * Creates a plain object from a Transaction message. Also converts values to other types if specified.
     * @param message Transaction
     * @param [options] Conversion options
     * @returns Plain object
     */
    public static toObject(message: Transaction, options?: $protobuf.IConversionOptions): { [k: string]: any };

    /**
     * Converts this Transaction to JSON.
     * @returns JSON object
     */
    public toJSON(): { [k: string]: any };
}

export namespace Transaction {

    /** Type enum. */
    enum Type {
        UNDEFINED = 0,
        INCOME = 1,
        EXPENSE = 2,
        TRANSFER = 3
    }
}